+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Version 0.1.1 (2019-04-24)
 
 * feature: account for loops in beta re-writes .. now the identity test works
 * enhancement: minor fixes to examples and drop redundant examples
+* feature: implement pair/type T/FRIN/FROUT with m/l/r in Utils.type_side_to_inout
 
 
 Version 0.1.0 (2019-04-24)